action_recognition_module/recognition-model/requirements_ar_venv.txt
antlr4-python3-runtime==4.9.2
asn1crypto==1.5.1
av==12.0.0
black==21.4b2
chardet==5.2.0
cloudpickle==3.0.0
contourpy==1.1.1
decord==0.6.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@9258799e4e72786edd67940872e0ed2c4387aac5
editdistance==0.6.0
einops @ git+https://github.com/arogozhnikov/einops.git@309337ddc9e9ae90e463dcba2dc4905b347eda37
funcsigs==1.0.2
future==1.0.0
fvcore==0.1.5.post20221221
huggingface-hub==0.23.0
hydra-core==1.3.2
importlib_metadata==8.0.0
imutils==0.5.4
iopath==0.1.9
jsonpatch==1.33
llvmlite==0.41.1
lmdb==1.2.1
-e git+https://github.com/microsoft/GLIP.git@9dda9558c1ef59bb6cdc8e896e2bcab775a68ff0#egg=maskrcnn_benchmark
matplotlib==3.7.5
mypy-extensions==1.0.0
numba==0.58.1
numexpr==2.8.6
numpy==1.24.4
omegaconf==2.3.0
pandas==1.3.5
parameterized==0.9.0
pathspec==0.12.1
pillow==10.4.0
portalocker==2.10.1
psutil==5.9.0
pyaml==21.10.1
pycocotools==2.0.7
pyDeprecate==0.3.1
pydot==2.0.0
pytorch-lightning==1.5.6
pytorchvideo==0.1.5
regex==2024.5.15
scikit-learn==0.24.2
scikit-optimize==0.9.0
scipy==1.10.1
submitit==1.3.3
timm==0.4.12
toml==0.10.2
torchnet==0.0.4
typing_extensions==4.11.0
visdom==0.2.4
xgboost==1.7.1
yacs==0.1.8